+++ lib/widgets/stateful_widget.dart
@@ -92,7 +92,7 @@ abstract class AdharaState<T extends StatefulWidget> extends State<T> {
     r.eventHandler.unregister(tag, eventName);
   }
 
-  AdharaEvent trigger(String eventName, dynamic data) {
+  Future<AdharaEvent> trigger(String eventName, dynamic data) {
     return r.eventHandler.trigger(eventName, data, tag);
   }