+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.1.68] - 7th December, 2018
+
+* Introducing `setData(key,value)` and `getData(key)` in Adhara event
+* Event emitters can await for handler execution and get handler return value by using event.getData(tag)
+
 ## [0.1.66] - 24th November, 2018
 
 * loading indicator update